requirements.txt
kaggle
pyYAML
python-box
ensure
notebook
pandas


##-e .